+++ lib/appmap/middleware/remote_recording.rb
@@ -9,12 +9,11 @@ module AppMap
         require 'appmap/command/record'
         require 'appmap/command/upload'
         require 'appmap/trace/tracer'
-        require 'appmap/config'
         require 'json'
 
         @app = app
-        @features = AppMap.inspect(config)
-        @functions = @features.map(&:collect_functions).flatten
+        @config = AppMap.configure
+        AppMap::Hook.hook(@config)
       end
 
       def event_loop
@@ -32,7 +31,7 @@ module AppMap
         return [ false, 'Recording is already in progress' ] if @tracer
 
         @events = []
-        @tracer = AppMap::Trace.tracers.trace(@functions)
+        @tracer = AppMap::Trace.tracers.trace
         @event_thread = Thread.new { event_loop }
         @event_thread.abort_on_exception = true
 
@@ -73,7 +72,11 @@ module AppMap
           name: 'remote_recording'
         }
 
-        response = JSON.generate(version: AppMap::APPMAP_FORMAT_VERSION, classMap: @features, metadata: metadata, events: @events)
+        response = JSON.generate \
+          version: AppMap::APPMAP_FORMAT_VERSION,
+          classMap: AppMap.class_map(@events),
+          metadata: metadata,
+          events: @events
 
         [ true, response ]
       end
@@ -111,10 +114,6 @@ module AppMap
 
       def html_response?(headers)
         headers['Content-Type'] && headers['Content-Type'] =~ /html/
-      end
-
-      def config
-        @config ||= AppMap::Config.load_from_file 'appmap.yml'
       end
 
       def recording?