+++ spec/fixtures/rails4_users_app/app/controllers/api/users_controller.rb
@@ -6,7 +6,7 @@ module Api
     end
 
     def create
-      @user = create_user(params.slice(:login).to_unsafe_h)
+      @user = build_user(params.slice(:login).to_unsafe_h)
       unless @user.valid?
         error = {
           code: 'invalid',
@@ -18,9 +18,9 @@ module Api
       end
       @user.save
       render json: @user, status: :created
-    end 
+    end
 
-    def create_user(params)
+    def build_user(params)
       User.new(params)
     end
   end