+++ test/fixtures/active_record_like/active_record/aggregations.rb
@@ -1,4 +1,0 @@
-module ActiveRecord
-  class Aggregations
-  end
-end