+++ test/fixtures/active_record_like/active_record/association.rb
@@ -1,4 +1,0 @@
-module ActiveRecord
-  module Associations
-  end
-end