+++ test/fixtures/active_record_like/active_record/associations/join_dependency.rb
@@ -1,6 +1,0 @@
-module ActiveRecord
-  module Associations
-    module JoinDependency
-    end
-  end
-end