+++ test/fixtures/active_record_like/active_record/caps/caps.rb
@@ -1,4 +1,0 @@
-module ActiveRecord
-  module CAPS
-  end
-end