+++ test/fixtures/includes_excludes/lib/a/a_2.rb
@@ -1,6 +1,0 @@
-module A
-  class A2
-    def action
-    end
-  end
-end