+++ test/fixtures/inspect_multiple_subdirs/module_a/class_a.rb
@@ -1,5 +1,0 @@
-module ModuleA
-  # @appmap
-  class ClassA
-  end
-end